+++ lib/searchkick/index.rb
@@ -123,7 +123,7 @@ module Searchkick
 
     def similar_record(record, options = {})
       like_text = retrieve(record).to_hash
-        .keep_if { |k, v| !options[:fields] || options[:fields].map(&:to_s).include?(k) }
+        .keep_if { |k, _| !options[:fields] || options[:fields].map(&:to_s).include?(k) }
         .values.compact.join(" ")
 
       # TODO deep merge method
@@ -141,9 +141,7 @@ module Searchkick
 
     def search_model(searchkick_klass, term = nil, options = {}, &block)
       query = Searchkick::Query.new(searchkick_klass, term, options)
-      if block
-        block.call(query.body)
-      end
+      block.call(query.body) if block
       if options[:execute] == false
         query
       else
@@ -372,6 +370,9 @@ module Searchkick
 
         # synonyms
         synonyms = options[:synonyms] || []
+
+        synonyms = synonyms.call if synonyms.respond_to?(:call)
+
         if synonyms.any?
           settings[:analysis][:filter][:searchkick_synonym] = {
             type: "synonym",
@@ -402,7 +403,7 @@ module Searchkick
         end
 
         if options[:special_characters] == false
-          settings[:analysis][:analyzer].each do |analyzer, analyzer_settings|
+          settings[:analysis][:analyzer].each do |_, analyzer_settings|
             analyzer_settings[:filter].reject! { |f| f == "asciifolding" }
           end
         end
@@ -410,8 +411,8 @@ module Searchkick
         mapping = {}
 
         # conversions
-        if options[:conversions]
-          mapping[:conversions] = {
+        if (conversions_field = options[:conversions])
+          mapping[conversions_field] = {
             type: "nested",
             properties: {
               query: {type: "string", analyzer: "searchkick_keyword"},
@@ -584,6 +585,5 @@ module Searchkick
         obj
       end
     end
-
   end
 end