+++ test/reindex_v2_job_test.rb
@@ -1,7 +1,6 @@
 require_relative "test_helper"
 
-class TestReindexV2Job < Minitest::Test
-
+class ReindexV2JobTest < Minitest::Test
   def setup
     skip unless defined?(ActiveJob)
     super
@@ -30,5 +29,4 @@ class TestReindexV2Job < Minitest::Test
     Product.searchkick_index.refresh
     assert_search "*", []
   end
-
 end